Updated dependencies for client build
