+++ cmd/gollama/gpuinfo.go
@@ -1,3 +1,5 @@
+//go:build !client
+
 package main
 
 import (
