+++ cmd/gollama/main.go
@@ -55,7 +55,6 @@ type Globals struct {
 
 type CLI struct {
 	Globals
-	GpuInfo GpuInfoCmd `cmd:"" name:"gpuinfo" help:"Show GPU information"`
 	ModelCommands
 	CompletionCommands
 	EmbedCommands
